+++ database/migrations/2020_02_02_000001_update_users_settings_table.php
@@ -8,17 +8,21 @@ class UpdateUsersSettingsTable extends Migration
 {
     public function up()
     {
-        Schema::table('users_settings', function (Blueprint $table) {
-            $table->boolean('welcome_flow_enabled')->default(1);
-            $table->tinyInteger('welcome_flow_step')->default(1);
-        });
+        if(Schema::hasTable('users_settings')) {
+            Schema::table('users_settings', function (Blueprint $table) {
+                $table->boolean('welcome_flow_enabled')->default(1);
+                $table->tinyInteger('welcome_flow_step')->default(1);
+            });
+        }
     }
 
     public function down()
     {
-        Schema::table('users_settings', function (Blueprint $table) {
-            $table->dropColumn('welcome_flow_enabled');
-            $table->dropColumn('welcome_flow_step');
-        });
+        if(Schema::hasTable('users_settings')) {
+            Schema::table('users_settings', function (Blueprint $table) {
+                $table->dropColumn('welcome_flow_enabled');
+                $table->dropColumn('welcome_flow_step');
+            });
+        }
     }
 }
